tests: Provide proper -Wl switch to use ld-2.30.so instead of old ld-2.28.so
glibc -master commit 1bdda52fe92fd01b424cd6fbb63e3df96a95015c
Date:   Fri Nov 29 10:44:59 2019 -0300

    elf: Move vDSO setup to rtld (BZ#24967)

Changed the layout of
struct rtld_global_ro @ sysdeps/generic/ldsodefs.h

+# define PROCINFO_DECL
+# ifndef PROCINFO_CLASS
+#  define PROCINFO_CLASS EXTERN
+# endif
+# include <dl-vdso-setup.c>

The old ld-2.28.so used when test_{ny}2038 was run was not aware
of this change and caused SIGSEGV error (as layout mismatches).

The glibc project only support matching versions of libc.so and ld.so
This particular patch writes to ELF binary exact patch (and hence the
version) of the ld.so version to use.

Signed-off-by: Lukasz Majewski <lukma@denx.de>