glibc: Adjust the meta-y2038 layer to use glibc from -master branch
Until this change the glibc, which was supporting y2038, was used only
as DUT (installed in /opt 'sandbox'). The meta-y2038 was the 'thud' YPRR,
which was using the 2.28 glibc.

After this change, the newest glibc (pre 2.34 release) with 64-bit time
support on 32 bit ports is used.

The README file has been adjusted as well.

Note:
-----

Following changes had to be done to allow building the meta-y2038 with
pre-release version of glibc 2.34 with -master Yocto project (hardknott):

1. Backport texinfo to 6.7

2. Adjust util-linux to use close_range() syscall correctly

3. I had to revert glibc commit which caused build breaks on gcc-runtime
libgomp compilation

All above issues shall be fixed when first Yocto release uses 2.34 glibc
as its base version.

Signed-off-by: Lukasz Majewski <lukma@denx.de>