qemu: arm: Remove support for MACHINE=y2038arm and use qemuarm instead
The 'y2038arm' machine was devised on the very beginning of Y2038
development process. It was using the Versatile board with 2 Cortex-A9
cores.

After this change the meta-y2038 would solely use the 'qemuarm' machine,
which is actively maintained by Poky/Yocto.

In the contemporary poky, the default QEMU emulated ARM machine is more
powerfull 4 core Cortex-A15 board.

Signed-off-by: Lukasz Majewski <lukma@denx.de>