conf: Disable image-prelink class (to fix glibc's ARM ld-linux-armhf.so.3 issue)
The image-prelink.bbclass causes all in-rootfs binaries and libraries
being tuned (i.e. the loading addresses are assinged beforehand) to
improve loading (and startup) time.

This operation is only done once, during creating the rootfs.
Disable it to avoid issue with ld-linux-armhf.so.3 relocation to kernel
space area (glibc's BZ #28293)

Signed-off-by: Lukasz Majewski <lukma@denx.de>